Use the dap-js SDK
